Refactor MCP validate and DTO tools
