+++ crates/rulemorph_mcp/src/tools/mod.rs
@@ -1,5 +1,7 @@
 pub(crate) mod analyze_input;
+pub(crate) mod generate_dto;
 pub(crate) mod generate_rules_from_base;
 pub(crate) mod generate_rules_from_dto;
 pub(crate) mod list_ops;
 pub(crate) mod transform;
+pub(crate) mod validate;
