+++ crates/rulemorph_mcp/src/tools/validate.rs
@@ -1,0 +1,68 @@
+use rulemorph::validate_rule_file_with_source;
+use serde_json::{Map, Value, json};
+
+use crate::args::get_optional_string;
+use crate::diagnostics::{
+    collect_rule_warnings, rule_warnings_to_json, validation_errors_to_values,
+};
+use crate::errors::CallError;
+use crate::rule_source::load_rule_from_source;
+
+pub(crate) fn run_validate_rules_tool(args: &Map<String, Value>) -> Result<Value, CallError> {
+    let rules_path = get_optional_string(args, "rules_path").map_err(CallError::InvalidParams)?;
+    let rules_text = get_optional_string(args, "rules_text").map_err(CallError::InvalidParams)?;
+    let rules_format =
+        get_optional_string(args, "rules_format").map_err(CallError::InvalidParams)?;
+
+    let rule_source_count = rules_path.is_some() as u8 + rules_text.is_some() as u8;
+    if rule_source_count == 0 {
+        return Err(CallError::InvalidParams(
+            "rules_path or rules_text is required".to_string(),
+        ));
+    }
+    if rule_source_count > 1 {
+        return Err(CallError::InvalidParams(
+            "rules_path and rules_text are mutually exclusive".to_string(),
+        ));
+    }
+
+    let (rule, yaml, _) = load_rule_from_source(
+        rules_path.as_deref(),
+        rules_text.as_deref(),
+        rules_format.as_deref(),
+    )?;
+    match validate_rule_file_with_source(&rule, &yaml) {
+        Ok(_) => {
+            let warnings = collect_rule_warnings(&rule);
+            let mut result = json!({
+                "content": [
+                    {
+                        "type": "text",
+                        "text": "ok"
+                    }
+                ]
+            });
+            if !warnings.is_empty() {
+                result["meta"] = json!({
+                    "warnings": rule_warnings_to_json(&warnings)
+                });
+            }
+            Ok(result)
+        }
+        Err(errors) => {
+            let error_values = validation_errors_to_values(&errors);
+            Ok(json!({
+                "content": [
+                    {
+                        "type": "text",
+                        "text": "validation failed"
+                    }
+                ],
+                "isError": true,
+                "meta": {
+                    "errors": error_values
+                }
+            }))
+        }
+    }
+}
